Moves AddressLookupTable into SDK as well.
