+++ cli/src/address_lookup_table.rs
@@ -1,7 +1,6 @@
 use {
     crate::cli::{CliCommand, CliCommandInfo, CliConfig, CliError, ProcessResult},
     clap::{App, AppSettings, Arg, ArgMatches, SubCommand},
-    solana_address_lookup_table_program::state::AddressLookupTable,
     solana_clap_utils::{self, input_parsers::*, input_validators::*, keypair::*},
     solana_cli_output::{CliAddressLookupTable, CliAddressLookupTableCreated, CliSignature},
     solana_remote_wallet::remote_wallet::RemoteWalletManager,
@@ -15,6 +14,7 @@ use {
                 close_lookup_table, create_lookup_table, create_lookup_table_signed,
                 deactivate_lookup_table, extend_lookup_table, freeze_lookup_table,
             },
+            state::AddressLookupTable,
         },
         clock::Clock,
         commitment_config::CommitmentConfig,
