+++ programs/address-lookup-table-tests/tests/deactivate_lookup_table_ix.rs
@@ -3,10 +3,9 @@ use {
     common::{
         add_lookup_table_account, assert_ix_error, new_address_lookup_table, setup_test_context,
     },
-    solana_address_lookup_table_program::state::AddressLookupTable,
     solana_program_test::*,
     solana_sdk::{
-        address_lookup_table::instruction::deactivate_lookup_table,
+        address_lookup_table::{instruction::deactivate_lookup_table, state::AddressLookupTable},
         instruction::InstructionError,
         pubkey::Pubkey,
         signature::{Keypair, Signer},
