+++ programs/address-lookup-table/src/lib.rs
@@ -2,6 +2,4 @@
 #![cfg_attr(RUSTC_WITH_SPECIALIZATION, feature(specialization))]
 #![cfg_attr(RUSTC_NEEDS_PROC_MACRO_HYGIENE, feature(proc_macro_hygiene))]
 
-#[cfg(not(target_os = "solana"))]
 pub mod processor;
-pub mod state;
